CP-28247: reproduce clustering clean shutdown failure
Actually check for crashdumps after runnings tests.
Reduce timeouts on clustering to trigger this crash more easily.
Skip lifecycle operations on master because they interrupt the API
connection and prevent further testing.
Run operations on all hosts, even if we tried them previously, this is
needed for reboot to be tested at all because it only shows up when a
host is disabled.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>