Do not start XAPI in clustering component tests
Generating a new SSL certificate causes `xapi-nbd` to get started due to
path activation, which also starts xapi and a lot of other daemons.
We want to test clustering in isolation, so do ensure we do not start
xapi-nbd via path activation by masking the service.

Also update the branch, the feature branch got deleted, it is all in
master now.

Change N to 16 to speed up create operations, we do not support beyond
16 yet.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>