CP-29398: ensure rsync correctly uploads files
Previously it happened from time to time that rsync did not completely
upload the xapi-clusterd binary, and this lead to checksum mismatch
errors in the xapi-clusterd component tests.

I've added rsync flags --verbose --itemize-changes for debugging,
--checksum for avoiding file timestamp errors, and --whole-file to rule
out any issue with the delta copying algorithm. This combination of
flags seems to have fixed the issue. I have found that --whole-file is
necessary, having --checksum is not enough.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>